refactor: 사기 유형 특징 분리 및 적절한 요건에 렌더링되도록 수정
